backend/requirements.txt
crewai==0.70.1
crewai-tools==0.12.1
firecrawl-py==1.5.0
supabase==2.9.0
python-dotenv==1.0.1
pydantic==2.9.2
pypdf==5.1.0
playwright==1.48.0
thefuzz==0.22.1
python-Levenshtein==0.26.0
langchain==0.3.7
langchain-openai==0.2.8
openai==1.54.4
requests==2.32.3
